feat(personhog-replica): use shutdown status for readiness endpoint
