+++ rust/capture/src/metrics_middleware.rs
@@ -1,5 +1,5 @@
 use std::{
-    sync::atomic::{AtomicU8, AtomicUsize, Ordering},
+    sync::atomic::{AtomicUsize, Ordering},
     time::{Duration, Instant},
 };
 
@@ -13,53 +13,11 @@ use axum::{
 };
 use metrics::gauge;
 
+// Re-exporting from health crate for backwards compatibility
+pub use health::{get_shutdown_status, set_shutdown_status, ShutdownStatus};
+
 // Global atomic counter for active connections
 static ACTIVE_CONNECTIONS: AtomicUsize = AtomicUsize::new(0);
-
-// Shutdown status state machine
-#[derive(Clone, Copy, PartialEq, Eq, Debug)]
-#[repr(u8)]
-pub enum ShutdownStatus {
-    Unknown = 0,
-    Running = 1,
-    Prestop = 2,
-    Terminating = 3,
-    Completed = 4,
-}
-
-impl ShutdownStatus {
-    pub fn as_str(self) -> &'static str {
-        match self {
-            Self::Unknown => "unknown",
-            Self::Running => "running",
-            Self::Prestop => "prestop",
-            Self::Terminating => "terminating",
-            Self::Completed => "completed",
-        }
-    }
-}
-
-impl From<u8> for ShutdownStatus {
-    fn from(v: u8) -> Self {
-        match v {
-            1 => Self::Running,
-            2 => Self::Prestop,
-            3 => Self::Terminating,
-            4 => Self::Completed,
-            _ => Self::Unknown,
-        }
-    }
-}
-
-static SHUTDOWN_STATUS: AtomicU8 = AtomicU8::new(ShutdownStatus::Running as u8);
-
-pub fn set_shutdown_status(status: ShutdownStatus) {
-    SHUTDOWN_STATUS.store(status as u8, Ordering::Relaxed);
-}
-
-pub fn get_shutdown_status() -> ShutdownStatus {
-    SHUTDOWN_STATUS.load(Ordering::Relaxed).into()
-}
 
 // Guard to ensure connection count is decremented even on panic
 struct ConnectionGuard;
