+++ rust/personhog-replica/Cargo.toml
@@ -9,6 +9,7 @@ personhog-proto = { path = "../personhog-proto" }
 common-database = { path = "../common/database" }
 common-metrics = { path = "../common/metrics" }
 common-alloc = { path = "../common/alloc" }
+health = { path = "../common/health" }
 
 async-trait = { workspace = true }
 axum = { workspace = true }
